+++ example/android/app/src/main/kotlin/com/sandromaglione/repocase/example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.sandromaglione.repo.case.example
+package com.sandromaglione.repocase.example
 
 import io.flutter.embedding.android.FlutterActivity