+++ repo_case_generator/test/repo_case_generator_test.dart
@@ -1,4 +1,3 @@
-// @dart=2.9
 @TestOn('vm')
 
 import 'package:test/test.dart';