+++ repo_case_generator/test/src/generator_tests.dart
@@ -1,4 +1,3 @@
-// @dart=2.9
 import 'package:repo_case/repo_case.dart';
 import 'package:source_gen_test/annotations.dart';